pyproject.toml
[project]
name = "drf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "pytest-django>=4.11.1",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]
